syntax and cleanup
