+++ avakas/flavors/__init__.py
@@ -2,7 +2,7 @@
 Avakas Built-In Project Flavors
 """
 
-from .base import AvakasLegacy, AvakasGitProject
+from .base import AvakasLegacy
 from .ansible import AvakasAnsibleProject
 from .chef import AvakasChefProject
 from .erlang import AvakasErlangProject
@@ -10,7 +10,6 @@ from .node import AvakasNodeProject
 
 __all__ = [
     'AvakasLegacy',
-    'AvakasGitProject',
     'AvakasAnsibleProject',
     'AvakasChefProject',
     'AvakasErlangProject',
