+++ avakas/utils.py
@@ -3,9 +3,26 @@ Avakas Utility Functions
 """
 
 import re
+import sys
+import os
 from functools import cmp_to_key
+import contextlib
 
 from semantic_version import compare
+
+
+@contextlib.contextmanager
+def stdout_redirect():
+    """ Forcefully redirect stdout to stderr """
+    # http://marc-abramowitz.com/archives/2013/07/19/python-context-manager-for-redirected-stdout-and-stderr/
+    try:
+        oldstdchannel = os.dup(sys.stdout.fileno())
+        os.dup2(sys.stderr.fileno(), sys.stdout.fileno())
+
+        yield
+    finally:
+        if oldstdchannel is not None:
+            os.dup2(oldstdchannel, sys.stdout.fileno())
 
 
 def match_and_rewrite_lines(pattern, file_body, version):
